requirements.txt
wxPython==4.1.1
Pillow==8.3.2
pycryptodome==3.10.1